ecommerce-api/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
sqlalchemy==2.0.12
pymysql==1.0.3
pydantic==1.10.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.0.0
pydantic-settings==0.2.5


